pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.60.0",
    "pyyaml>=6.0.3",
    "sift-stack-py>=0.9.6",
    "streamlit>=1.53.0",
]
